+++ src/main/java/graphql/schema/diff/DiffCategory.java
@@ -1,4 +1,4 @@
-package graphql.diff;
+package graphql.schema.diff;
 
 import graphql.PublicApi;
